Update for Hexo 3
